+++ src/mongo/db/pipeline/document_source_group_test.cpp
@@ -396,7 +396,7 @@ protected:
         }
     }
 
-    virtual boost::optional<size_t> getMaxMemoryUsageBytes() {
+    virtual boost::optional<int64_t> getMaxMemoryUsageBytes() {
         return boost::none;
     }
 
@@ -1096,7 +1096,7 @@ private:
         }
     }
 
-    boost::optional<size_t> getMaxMemoryUsageBytes() final {
+    boost::optional<int64_t> getMaxMemoryUsageBytes() final {
         return 10 * kBigStringSize;
     }
 
@@ -1140,7 +1140,7 @@ private:
                           << "$monotonicIdFields" << BSON_ARRAY("x"));
     }
 
-    boost::optional<size_t> getMaxMemoryUsageBytes() final {
+    boost::optional<int64_t> getMaxMemoryUsageBytes() final {
         return 10 * kBigStringSize;
     }
 
@@ -1189,7 +1189,7 @@ private:
                     << "$monotonicIdFields" << BSON_ARRAY("_id"));
     }
 
-    boost::optional<size_t> getMaxMemoryUsageBytes() final {
+    boost::optional<int64_t> getMaxMemoryUsageBytes() final {
         return (25 * kBigStringSize) / 10;
     }
 
@@ -1236,7 +1236,7 @@ private:
                           << "$monotonicIdFields" << BSON_ARRAY("x"));
     }
 
-    boost::optional<size_t> getMaxMemoryUsageBytes() final {
+    boost::optional<int64_t> getMaxMemoryUsageBytes() final {
         return 10 * kBigStringSize;
     }
 
@@ -1273,7 +1273,7 @@ private:
             "{_id: {x: '$x', y: '$y'}, sum: {$sum: '$z'}, $monotonicIdFields: ['x', 'y']}");
     }
 
-    boost::optional<size_t> getMaxMemoryUsageBytes() final {
+    boost::optional<int64_t> getMaxMemoryUsageBytes() final {
         return 10 * kBigStringSize;
     }
